+++ CHANGELOG.md
@@ -10,6 +10,16 @@ As a bonus, we may try to document porting steps for users of this library that 
 
 ## [Unreleased]
 
+### Changed
+- azaSampler interface
+	- Uses ADSR envelopes
+	- Can now have multiple instances of sounds with independent pitch and gain
+	- Pitch and gain follows targets in a predictable, linear fashion
+
+### Added
+- ADSR envelopes
+- azaFollowerLinear for linearly following target values at a given rate
+
 ## [v0.2.1](https://github.com/Equivocatorrr/AzAudio/releases/tag/v0.2.1)
 
 ### Added